ETCM-196: Separate Payload from Packet.
